refactored timeline
