+++ src/components/home-page/background/fullTimeLine.module.css
@@ -1,5 +1,6 @@
 @import '../../layout/variables.css';
 .main{
+  padding-top:36px;
   h1{
     padding-left:18px;
     background-color:var(--dark_green);
