+++ src/components/home-page/introduction/Introduction.js
@@ -20,6 +20,8 @@ export default function headShotWithText(){
     setNavState({ display: textToDisplay()});
   };
 
+  const iconColor=styles.icon_color;
+  console.log('icon color', iconColor);
 
   const iconsWithToggleText = [...desktopContactIcons, {
     icon: <FaEnvelope/>,
@@ -48,12 +50,12 @@ export default function headShotWithText(){
         <section className={styles.container_text}>
           {navState.display === 'about_me' ? <h3 className={styles.desktop_title}>{title}</h3> : <h3 className={styles.desktop_title}>{contactMe.title}</h3>}
           <section className={styles.mobile_contact}>
-          <RenderIcon iconArray={iconsWithToggleText} direction='row' color='rgba(19, 32, 25, 1)'/>
+          <RenderIcon iconArray={iconsWithToggleText} direction='row' color={iconColor} className={styles.iconColor}/>
           </section>
           {navState.display === 'about_me' ? <p>{aboutMe}</p> : <p>{contactMe.content}</p>}
         </section>
       <section className={styles.desktop_contact}>
-        <RenderIcon iconArray={iconsWithToggleText} direction='column' color='rgba(19, 32, 25, 1)'/>
+        <RenderIcon iconArray={iconsWithToggleText} direction='column' color={iconColor} className={styles.iconColor}/>
       </section>
     </section>
   )
