+++ src/components/render-icon/iconWithContent.module.css
@@ -1,4 +1,7 @@
+@import '../layout//variables.css';
+
 .icon_with_content{
   display:flex;
   justify-content: center;
+  color: var(--dark_green)
 }
